+++ wger/manager/api/views.py
@@ -35,10 +35,12 @@ from wger.manager.api.serializers import (
     ScheduleStepSerializer,
     SetSerializer,
     SettingSerializer,
+    WorkoutAndTemplateSerializer,
     WorkoutCanonicalFormSerializer,
     WorkoutLogSerializer,
     WorkoutSerializer,
     WorkoutSessionSerializer,
+    WorkoutTemplateSerializer,
 )
 from wger.manager.models import (
     Day,
@@ -112,6 +114,72 @@ class WorkoutViewSet(viewsets.ModelViewSet):
         for key, values in entry_logs.items():
             serialized_logs[str(key)] = [WorkoutLogSerializer(entry).data for entry in values]
         return Response({'chart_data': json.loads(chart_data), 'logs': serialized_logs})
+
+
+class UserWorkoutTemplateViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint for workout objects
+    """
+    serializer_class = WorkoutTemplateSerializer
+    is_private = True
+    ordering_fields = '__all__'
+    filterset_fields = ('name', 'description', 'creation_date')
+
+    def get_queryset(self):
+        """
+        Only allow access to appropriate objects
+        """
+        return Workout.templates.filter(user=self.request.user)
+
+    def perform_create(self, serializer):
+        """
+        Set the owner
+        """
+        serializer.save(user=self.request.user)
+
+
+class PublicWorkoutTemplateViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint for public workout templates objects
+    """
+    serializer_class = WorkoutSerializer
+    is_private = True
+    ordering_fields = '__all__'
+    filterset_fields = ('name', 'description', 'creation_date')
+
+    def get_queryset(self):
+        """
+        Only allow access to appropriate objects
+        """
+        return Workout.templates.filter(is_public=True)
+
+    def perform_create(self, serializer):
+        """
+        Set the owner
+        """
+        serializer.save(user=self.request.user)
+
+
+class WorkoutAndTemplateViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint for workout objects
+    """
+    serializer_class = WorkoutAndTemplateSerializer
+    is_private = True
+    ordering_fields = '__all__'
+    filterset_fields = ('name', 'description', 'creation_date')
+
+    def get_queryset(self):
+        """
+        Only allow access to appropriate objects
+        """
+        return Workout.both.filter(user=self.request.user)
+
+    def perform_create(self, serializer):
+        """
+        Set the owner
+        """
+        serializer.save(user=self.request.user)
 
 
 class WorkoutSessionViewSet(WgerOwnerObjectModelViewSet):