+++ wger/urls.py
@@ -58,6 +58,12 @@ router.register(r'day', manager_api_views.DayViewSet, basename='day')
 router.register(r'set', manager_api_views.SetViewSet, basename='Set')
 router.register(r'setting', manager_api_views.SettingViewSet, basename='Setting')
 router.register(r'workout', manager_api_views.WorkoutViewSet, basename='workout')
+router.register(r'templates', manager_api_views.UserWorkoutTemplateViewSet, basename='templates')
+router.register(
+    r'public-templates',
+    manager_api_views.PublicWorkoutTemplateViewSet,
+    basename='public-templates',
+)
 router.register(
     r'workoutsession', manager_api_views.WorkoutSessionViewSet, basename='workoutsession'
 )